Vulkan Fixed uninitialized variables 10/31/2019 | 19w44a3 | 0.5.59
& made the config class usable for other files than Engine.cfg

+++ TRAP/src/Graphics/Buffers/BufferLayout.h
@@ -25,7 +25,7 @@ namespace TRAP::Graphics
 	{
 		std::string Name;
 		ShaderDataType Type{ShaderDataType::None};
-		intptr_t Size{};
+		uint32_t Size{};
 		uint32_t Offset{};
 		bool Normalized{};
